feat: messages in seperate files

+++ src/modules/ServerAdministrator/DonationModule.ts
@@ -1,7 +1,7 @@
 import { command, CommandCategories } from "@utils/commands";
 import { monitor } from "@utils/monitor";
 import inhibitors from "@utils/inhibitors";
-import { commandDescriptions, errorMessage, strings } from "@utils/messages";
+import strings, { commandDescriptions } from "@utils/messages";
 import { Module } from "@utils/modules";
 import * as Arguments from "@utils/arguments";
 import Discord from "discord.js";
@@ -14,7 +14,7 @@ export default class DonationModule extends Module {
     if (![env.ROLES.HYPERION, env.ROLES.EVOCATION_MIRACULUM].includes(role.id)) return;
     const auditLogs = await newMember.guild.fetchAuditLogs({ type: "MEMBER_ROLE_UPDATE" });
     const firstEntry = auditLogs.entries.first();
-    if (!(firstEntry.changes[0].key === "$add" && [this.client.user.id].includes(firstEntry.executor.id))) { newMember.roles.remove(role, strings.modules.donations.auditLogDonationRoleAdd); }
+    if (!(firstEntry.changes[0].key === "$add" && [this.client.user.id].includes(firstEntry.executor.id))) { newMember.roles.remove(role, strings.modules.administrator.donations.auditLogDonationRoleAdd); }
   }
 
   @command({
@@ -23,15 +23,15 @@ export default class DonationModule extends Module {
   logdonation(msg: Discord.Message, member: Discord.GuildMember, item: string): void {
     msg.delete();
     if (member.user.bot) {
-      errorMessage(msg, strings.general.error(strings.modules.donations.commands.logdonationBotError));
+      strings.errors.errorMessage(msg, strings.errors.error(strings.modules.administrator.donations.commands.logdonationBotError));
       return;
     }
-    if (member.roles.cache.has(env.ROLES.HYPERION)) msg.channel.send(strings.general.success(strings.modules.donations.commands.logdonationAlreadyHyperion(member.user)), { allowedMentions: { roles: [], users: [] } }).then((m) => setTimeout(() => m.delete(), 5000));
+    if (member.roles.cache.has(env.ROLES.HYPERION)) msg.channel.send(strings.general.success(strings.modules.administrator.donations.commands.logdonationAlreadyHyperion(member.user)), { allowedMentions: { roles: [], users: [] } }).then((m) => setTimeout(() => m.delete(), 5000));
     else {
       member.roles.add(env.ROLES.HYPERION);
-      msg.channel.send(strings.general.success(strings.modules.donations.commands.logdonationNewHyperion(member.user)), { allowedMentions: { roles: [], users: [] } }).then((m) => setTimeout(() => m.delete(), 5000));
+      msg.channel.send(strings.general.success(strings.modules.administrator.donations.commands.logdonationNewHyperion(member.user)), { allowedMentions: { roles: [], users: [] } }).then((m) => setTimeout(() => m.delete(), 5000));
     }
-    (msg.guild.channels.resolve(env.CHANNELS.DONATION_LOG) as Discord.TextChannel).send(strings.modules.donations.commands.logdonationLogEntry(member.user, item, msg.author));
+    (msg.guild.channels.resolve(env.CHANNELS.DONATION_LOG) as Discord.TextChannel).send(strings.modules.administrator.donations.commands.logdonationLogEntry(member.user, item, msg.author));
   }
 
   @command({
@@ -39,6 +39,6 @@ export default class DonationModule extends Module {
   })
   async miraculum(msg: Discord.Message, member: Discord.GuildMember): Promise<void> {
     await member.roles.add(env.ROLES.EVOCATION_MIRACULUM);
-    msg.channel.send(strings.general.success(strings.modules.donations.commands.awardMiraculum(member.user)), { allowedMentions: { roles: [], users: [] } });
+    msg.channel.send(strings.general.success(strings.modules.administrator.donations.commands.awardMiraculum(member.user)), { allowedMentions: { roles: [], users: [] } });
   }
 }
